+++ lib/wed/input-trigger.ts
@@ -65,7 +65,8 @@ export class InputTrigger {
    */
   constructor(private readonly editor: Editor,
               private readonly selector: string) {
-    this.guiSelector = toGUISelector(this.selector);
+    this.guiSelector =
+      toGUISelector(this.selector, editor.mode.getAbsoluteNamespaceMappings());
 
     // This is a map of keys that are actually text keys to their handlers. This
     // map is in effect a submap of _key_to_handler. We want this for speed,